+++ tests/tests_integration/test_api/test_geospatial.py
@@ -2,6 +2,7 @@ import random
 import re
 import time
 import uuid
+from pathlib import Path
 
 import pytest
 
@@ -24,6 +25,8 @@ from cognite.client.exceptions import CogniteAPIError
 from tests.utils import set_request_limit
 
 FIXED_SRID = 121111 + random.randint(0, 1_000)
+
+GEOSPATIAL_TEST_RESOURCES = Path(__file__).parent / "geospatial_test_resources"
 
 
 @pytest.fixture()
@@ -640,10 +643,8 @@ class TestGeospatialAPI:
             raster_property_name="raster",
             raster_format="XYZ",
         )
-        raster_content = open(
-            "tests/tests_integration/test_api/geospatial_test_resources/raster-grid-example.xyz", "rb"
-        ).read()
-        assert res == raster_content
+        raster_content = (GEOSPATIAL_TEST_RESOURCES / "raster-grid-example.xyz").read_text()
+        assert res.decode(encoding="utf-8") == raster_content
 
         res = cognite_client.geospatial.get_raster(
             feature_type_external_id=test_feature_type.external_id,
@@ -652,10 +653,8 @@ class TestGeospatialAPI:
             raster_format="XYZ",
             raster_options={"DECIMAL_PRECISION": 5},
         )
-        raster_content = open(
-            "tests/tests_integration/test_api/geospatial_test_resources/raster-grid-5-decimal.xyz", "rb"
-        ).read()
-        assert res == raster_content
+        raster_content = (GEOSPATIAL_TEST_RESOURCES / "raster-grid-5-decimal.xyz").read_text()
+        assert res.decode(encoding="utf-8") == raster_content
 
     def test_get_raster_with_transformation(self, cognite_client, test_feature_type, test_feature_with_raster):
         res = cognite_client.geospatial.get_raster(
@@ -666,10 +665,8 @@ class TestGeospatialAPI:
             raster_srid=54030,
             allow_crs_transformation=True,
         )
-        raster_content = open(
-            "tests/tests_integration/test_api/geospatial_test_resources/raster-grid-54030-example.xyz", "rb"
-        ).read()
-        assert res == raster_content
+        raster_content = (GEOSPATIAL_TEST_RESOURCES / "raster-grid-54030-example.xyz").read_text()
+        assert res.decode(encoding="utf-8") == raster_content
 
     def test_retrieve_features_with_raster_property(self, cognite_client, test_feature_type, test_feature_with_raster):
         res = cognite_client.geospatial.retrieve_features(